DA#229: Refactor: Modularize Document commands
