+++ src/commands/documents/index.ts
@@ -1,0 +1,5 @@
+export * from './createDocument';
+export * from './getDocumentMetaData';
+export * from './openDocument';
+export * from './removeDocument';
+export * from './searchDocument';
